remove term "Grant of Additional Permission"
It's a little odd to use "Grant of Additional Permission" as a formal title when neither GPL v3 or the rest of this license uses that term. 

This pull request proposes replacing that phrase with simply "Text of Exception".

For consistency, may also want to consider titling the first part "Preamble", but that would require a lot of reformatting so not proposing it without agreement.