Restructure 2d paragraph to parallel Sec. 5(c)
This restructures the second paragraph of the permission to parallel Sec. 5(c).

Downside: it is less comprehensible to mere humans :( 
Upside: it is no longer basically a clone of the "purpose" sentence, and it is more clear to an expert lawyer trying to pair this with Sec. 5 of the license.